Modbus: Change types of arguments in modbus setters.

+++ src/Modbus/ModbusSetter.cpp
@@ -3,7 +3,7 @@
 #include "modbus.h"
 
 namespace Modbus {
-bool setHoldingModbus(const configuration::Configuration &modbusConf, int address, int value)
+bool setHoldingModbus(const configuration::Configuration &modbusConf, uint16_t address, uint16_t value)
 {
   auto modbusConnection = modbus(modbusConf.mNetworkConfiguration.mModbusTcpAddress, modbusConf.mNetworkConfiguration.mTCPPort);
   modbusConnection.modbus_set_slave_id(modbusConf.mNetworkConfiguration.mSlaveId);
@@ -12,7 +12,7 @@ bool setHoldingModbus(const configuration::Configuration &modbusConf, int addres
   return true;
 }
 
-bool setCoilModbus(const configuration::Configuration &modbusConf, int address, bool value)
+bool setCoilModbus(const configuration::Configuration &modbusConf, uint16_t address, bool value)
 {
   auto modbusConnection = modbus(modbusConf.mNetworkConfiguration.mModbusTcpAddress, modbusConf.mNetworkConfiguration.mTCPPort);
   modbusConnection.modbus_set_slave_id(modbusConf.mNetworkConfiguration.mSlaveId);
