+++ src/Modbus/ModbusSetter.h
@@ -4,5 +4,5 @@
 namespace Modbus {
 bool setModbus(const configuration::Configuration &modbusConf, int address, int value);
 
-bool setCoilModbus(const configuration::Configuration &modbusConf, int address, bool value);
+bool setCoilModbus(const configuration::Configuration &modbusConf, uint16_t address, bool value);
 }  // namespace Modbus
